Rename classes and fix project class/files structure
