+++ src/Command/Modx/Install.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace modmore\Gitify\Command\Modx;
 
 use modmore\Gitify\BaseCommand;
